backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
requests>=2.31.0
python-dotenv>=1.0.0
transformers>=4.41.0
torch>=2.3.0
Pillow>=10.3.0
numpy>=1.26.0
scikit-learn>=1.4.0
